+++ CHANGELOG.md
@@ -15,7 +15,10 @@ to [Semantic Versioning]. Full commit history is available in the
 - Add supervised module class {class}`scvi.module.base.SupervisedModuleClass`. {pr}`3237`.
 - Add get normalized function model property for any generative model {pr}`3238` and changed
     get_accessibility_estimates to get_normalized_accessibility, where needed.
-- Add support for using Lamin custom dataloaders with {class}`scvi.model.SCVI`, {pr}`2932`.
+- Add {class}`scvi.external.TOTALANVI`. {pr}`3259`.
+- Add Custom Dataloaders registry support, {pr}`2932`.
+- Add support for using Census and LaminAI custom dataloaders for {class}`scvi.model.SCVI`
+    and {class}`scvi.model.SCANVI`, {pr}`2932`.
 - Add Early stopping KL warmup steps. {pr}`3262`.
 - Add Minification option to {class}`~scvi.model.LinearSCVI` {pr}`3294`.
 
@@ -33,6 +36,8 @@ to [Semantic Versioning]. Full commit history is available in the
 - Updated resolVI to use rapids-singlecell. {pr}`3308`.
 
 #### Removed
+
+- Removed Jax version constraint for mrVI training. {pr}`3309`.
 
 ### 1.3.0 (2025-02-28)